Port to Window/VS2015

+++ source/fern/example/algorithm/test/raster_test.cc
@@ -16,7 +16,7 @@ BOOST_AUTO_TEST_SUITE(raster)
 
 BOOST_AUTO_TEST_CASE(constructor)
 {
-    size_t const cell_size = 5.0;
+    double const cell_size = 5.0;
     size_t const nr_rows = 600;
     size_t const nr_cols = 400;
     example::Raster<int32_t> raster1(cell_size, nr_rows, nr_cols);
